requirements.txt
python-telegram-bot==20.7 
yt-dlp==2023.12.30 
python-dotenv==1.0.0 
